+++ src/app/pages/BuyCrypto/hooks/use-tab-slug.ts
@@ -1,0 +1,12 @@
+import { useMemo } from 'react';
+
+import { useLocation } from 'lib/woozie';
+
+export const useTabSlug = () => {
+  const { search } = useLocation();
+  const tabSlug = useMemo(() => {
+    const usp = new URLSearchParams(search);
+    return usp.get('tab');
+  }, [search]);
+  return useMemo(() => tabSlug, [tabSlug]);
+};
